Refactored ("tags", "keywords") modules

+++ src/Domain/Modules/Tag/Services/tagMutations.services.ts
@@ -1,12 +1,11 @@
 import TagRepository from "./../Repository/tag.repository";
 import PostRepository from "../../Post/Repository/post.repository";
-import SeriesRepository from "../../Series/Repository/series.repository";
-import { CreateTagDTO, DeleteTagDTO, UpdateTagDTO, deleteTagsIfNotReferencedInOtherPostsOrSeriesDTO } from "../Types";
-import { InternalServerException, NotFoundException } from "../../../../Shared/Exceptions";
+import { CreateTagDTO, DeleteTagDTO, UpdateTagDTO } from "../Types";
+import { ForbiddenException, InternalServerException, NotFoundException } from "../../../../Shared/Exceptions";
 
 export default class TagMutationsServices {
 	public static async createTag(data: CreateTagDTO) {
-		const createdTag = await TagRepository.createOne({ ...data });
+		const createdTag = await TagRepository.createOne(data);
 
 		if (!createdTag) throw new InternalServerException("The Tag creation failed!");
 
@@ -23,35 +22,15 @@ export default class TagMutationsServices {
 	}
 
 	public static async deleteTag(data: DeleteTagDTO) {
-		/**
-		 * TODO: check other model!
-		 */
+		const tagReferencingPosts = await PostRepository.findMany({ tags: { $in: data._id } });
 
-		// if no other documents have this it as a reference, then
+		if (tagReferencingPosts.length !== 0)
+			throw new ForbiddenException("You need to delete the referencing posts first!");
+
 		const { deletedCount } = await TagRepository.deleteOne({ _id: data._id });
 
 		if (deletedCount === 0) throw new InternalServerException("The tag deletion process failed!");
 
 		return "The tag is deleted successfully!";
-	}
-
-	public static async deleteTagsIfNotReferencedInOtherPostsOrSeries(
-		data: deleteTagsIfNotReferencedInOtherPostsOrSeriesDTO
-	) {
-		const deletePromises = data.tags.map(async (tag) => {
-			const [foundPosts, foundSeries] = await Promise.all([
-				PostRepository.findMany({ tags: { $in: tag._id } }),
-				SeriesRepository.findMany({ tags: { $in: tag._id } }),
-			]);
-
-			const totalReferences = foundPosts.length + foundSeries.length;
-			console.log({ totalReferences });
-			console.log("from tags service");
-			if (totalReferences <= 3) {
-				await TagRepository.deleteOne({ _id: tag._id });
-			}
-		});
-
-		await Promise.all(deletePromises);
 	}
 }
